backend/requirements.txt
streamlit
requests
openai
plotly
fastapi
uvicorn
python-dotenv
 
# uvicorn main:app --reload
# python -m http.server 3000